requirements.txt
streamlit>=1.22.0
pandas>=1.5.3
numpy>=1.24.3
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.14.1
altair>=5.0.1
folium>=0.14.0
streamlit-folium>=0.11.1
pydeck>=0.8.0
snowflake-connector-python>=3.0.3
openpyxl>=3.1.2
requests>=2.30.0
python-dotenv>=1.0.0